Move charge confirmation sender endpoint to functionhost
